+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## [2.0.0]
- - Rename `parquetMetadata` to `parquetMetadataSync`
+ - Rename `parquetMetadataAsync` to `parquetMetadata`
  - Previous `parquetMetadata` is now `import { parquetMetadataSync } from 'hyparquet/src/metadata.js'`
 
 ## [1.23.3]